requirements.txt
apache-airflow
requests
pandas
sqlite3
